python/requirements.txt
fastapi[all]
google-cloud-storage
tensorflow
uvicorn
mltu==0.1.3
cv2
